Add files via upload
This dataset has been obtained from https://finance.yahoo.com/quote/DB/history?p=DB (Yahoo Finance) on 12th June 2020 and gives stock prices for Deutsche Bank stocks from 12/06/2019 to 12/06/2020. This will be used to create the csv file ..., which will in turn be used to create Figures (3a) and (3b) on page 22 of the report.