Add files via upload
This file gives the time to maturity, stock price (at the time) and actual price (at the given time) of the randomly selected Deutsche Bank call option  DB#F1920E950000. The maturity date of this call option was 12/06/2020, and we have given the aforementioned data for the time period 26/05/2020-12/06/2020. 
 The data for the actual call option price has been obtained from the bar chart found on MarketWatch (https://bigcharts.marketwatch.com/quickchart/quickchart.asp?symb=DB%23F1920E950000&insttype=&freq=1&show=&time=4).
The entries of the column 'Mat_time' have been calculated by finding the number of days between a given day (in the time period) and the maturity date, 12/06/2020. The entries of the column 'Stock price' have been obtained from DB_2.csv (given the corresponding times and dates). We use this file to code (in Python) for the plots in Figures (3a) and (3b).